eleuthia/requirements.txt
# Eleuthia Dependencies


imapclient>=2.3.1
python-telegram-bot[job-queue]>=20.0
APScheduler>=3.10.4
requests>=2.31.0
qrcode>=7.4.2
pillow>=10.0.0
flask>=3.0.0
libtmux
python-dotenv
pytz
litellm
pyotp
feedparser
python-dateutil
schedule
requests
numpy<2.0.0
chromadb==0.4.24
streamlit
langchain
langchain-community
langchain-chroma
langchain-litellm
langchain-text-splitters
pypdf